+++ Testing/Code/Review/itkLabelShapeKeepNObjectsImageFilterTest1.cxx
@@ -53,7 +53,7 @@ int itkLabelShapeKeepNObjectsImageFilterTest1(int argc, char * argv[])
   TEST_SET_GET_VALUE( BackgroundValue, KeepNObjects->GetBackgroundValue() );
 
   //testing get and set macros for Lambda 
-  double numberOfObjects = atof( argv[4] );
+  unsigned int numberOfObjects = atoi( argv[4] );
   KeepNObjects->SetNumberOfObjects( numberOfObjects );
   TEST_SET_GET_VALUE( numberOfObjects, KeepNObjects->GetNumberOfObjects() );
